Add Todo comment; Merge constants into privacy
