+++ packages/rum-core/src/index.ts
@@ -36,5 +36,4 @@ export * from './browser/htmlDomUtils'
 export * from './browser/polyfills'
 export { getSessionReplayUrl } from './domain/getSessionReplayUrl'
 export { isLongDataUrl, sanitizeDataUrl, MAX_ATTRIBUTE_VALUE_CHAR_LENGTH } from './domain/resource/resourceUtils'
-export * from './constants'
 export * from './domain/privacy'
